Add User class description and update README
Added docstring to the User class in users.py file to provide extensive details about its attributes, methods, and inherited classes or methods. The purpose of this addition is to enhance code readability and better inform contributors about the User class's functions and usage.

In the README, a section about SQLAlchemy and Alembic tools has been added to guide developers who aren't familiar with these tools. Additionally, a link to the Udemy course for these tools has been provided for more learning resources.