Update users.py
Refactoring of the User insertion process: now the insert operation will update the User entry if a conflict occurs (if a User with the provided user_id already exists). The updated attributes are username and full_name.
The use of .scalar_one() method from SQLAlchemy to return exactly one User model instance after the insert operation or raise an error if no result is found or if multiple results are found. This ensures the correctness of our function in situations where the insertion process modifies the state of an existing User.
Improved project code readability and database operation efficiency through more granular SQLAlchemy methods.